requirements.txt
streamlit
pyaudio
langchain
whisper
gtts
pygame
python-dotenv
langchain-groq
scipy
ffmpeg-python
